# Re-use reqwest client for better performance