--- a/src/cli.rs
+++ b/src/cli.rs
@@ -4,7 +4,7 @@ use std::io::{Read, Write};
 use std::process;
 use std::{thread, time};
 use trello::Renderable;
-use trello::{Card, CardContents, Client, TrelloError, TrelloObject};
+use trello::{Card, CardContents, TrelloClient, TrelloError, TrelloObject};
 
 pub fn multiselect_trello_object<T: TrelloObject + Renderable + PartialEq>(
     objects: &[T],
@@ -62,7 +62,7 @@ pub fn get_input(text: &str) -> Result<String, rustyline::error::ReadlineError> 
 ///
 /// This function will upload any changes written by the editor to Trello. This includes
 /// when the editor is not closed but content is saved.
-pub fn edit_card(client: &Client, card: &Card) -> Result<(), Box<dyn Error>> {
+pub fn edit_card(client: &TrelloClient, card: &Card) -> Result<(), Box<dyn Error>> {
     let mut file = tempfile::Builder::new().suffix(".md").tempfile()?;
     let editor_env = env::var("EDITOR").unwrap_or_else(|_| String::from("vi"));