Improve performance of subcommands by re-using reqwest::Client instead of using a new client each time a network request is performed as suggested in the documentation.

before
```
$ time tro show todo
tro show todo  0.06s user 0.02s system 5% cpu 1.369 total
```

after
```
$ time target/release/tro show todo
target/release/tro show todo  0.02s user 0.01s system 2% cpu 0.919 total
```